Change outrunTime to float and add small elo animation.
The configurable outrun time can now be set as a float, allowing for more precise outrun time settings.
Added a small animation for when the elo changes.

+++ TougePlugin/Race.cs
@@ -252,7 +252,7 @@ public class Race
         else if (LeaderSetLap && !FollowerSetLap)
         {
             // Make this time also configurable as outrun time.
-            int outrunTimer = _configuration.outrunTime * 1000;
+            int outrunTimer = (int)(_configuration.outrunTime * 1000f);
             _ = Task.Delay(outrunTimer).ContinueWith(_ => secondLapCompleted.TrySetResult(false));
         }
