+++ TougePlugin/TougeConfiguration.cs
@@ -28,7 +28,7 @@ public class TougeConfiguration : IValidateConfiguration<TougeConfigurationValid
     public bool isRollingStart = false;
 
     [YamlMember(Description = "Outrun timer in seconds. Chase car has to finish within this amount of time after the lead car crosses the finish line.")]
-    public int outrunTime = 3;
+    public float outrunTime = 3.0f;
 
     [YamlMember(Description = "Local database mode enabled. If disabled please provide database connection information.")]
     public bool isDbLocalMode = true;
